+++ application/src/main/java/common/bolts/transactional/tp/TPBolt_ts.java
@@ -50,8 +50,8 @@ public class TPBolt_ts extends TPBolt {
      * THIS IS ONLY USED BY TSTREAM.
      * IT CONSTRUCTS and POSTPONES TXNS.
      */
-    protected void PRE_TXN_PROCESS(long _bid, long timestamp) throws DatabaseException {
-        for (long i = _bid; i < _bid + combo_bid_size; i++) {
+    protected void PRE_TXN_PROCESS(double _bid, long timestamp) throws DatabaseException {
+        for (double i = _bid; i < _bid + combo_bid_size; i++) {
             TxnContext txnContext = new TxnContext(thread_Id, this.fid, i);
             LREvent event = (LREvent) input_event;
             (event).setTimestamp(timestamp);
