+++ application/src/main/java/common/param/ed/tc/TCEvent.java
@@ -14,13 +14,13 @@ public class TCEvent extends TxnEvent {
     private final SchemaRecordRef word_record = new SchemaRecordRef();
     public String[] tweetIDList;
     public boolean isBurst;
-    private final int myBid;
+    private final double myBid;
     private final int myPid;
     private final String my_bid_array;
     private final String my_partition_index;
     private final int my_number_of_partitions;
 
-    public TCEvent(int bid, int pid, String bid_array, String partition_index, int number_of_partitions,
+    public TCEvent(double bid, int pid, String bid_array, String partition_index, int number_of_partitions,
                    String wordID) {
         super(bid, pid, bid_array, partition_index, number_of_partitions);
         this.myBid = bid;
@@ -33,11 +33,11 @@ public class TCEvent extends TxnEvent {
         this.currWindow = computeCurrWindow(bid);
     }
 
-    private int computeCurrWindow(int bid) {
-        return bid / tweetWindowSize;
+    private int computeCurrWindow(double bid) {
+        return (int) bid / tweetWindowSize;
     }
 
-    public int getMyBid() {
+    public double getMyBid() {
         return myBid;
     }
     public int getMyPid() {
@@ -62,7 +62,7 @@ public class TCEvent extends TxnEvent {
     public SchemaRecordRef getWord_record() {return this.word_record;}
 
     public TCEvent cloneEvent() {
-        return new TCEvent((int) bid, pid, Arrays.toString(bid_array), Arrays.toString(partition_indexs), number_of_partitions,
+        return new TCEvent(bid, pid, Arrays.toString(bid_array), Arrays.toString(partition_indexs), number_of_partitions,
                 wordID);
     }
 }
